requirements.txt
asgiref==3.8.1
psycopg[binary]==3.3.2
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.5.1
resend==2.1.0
debugpy==1.8.20


